[Protocol3] Replace uint256 with uint for consistency (no logic change) (#307)

+++ packages/loopring_v3/contracts/iface/IExchange.sol
@@ -233,16 +233,16 @@ contract IExchange
     ///                      1st element of the array.
     /// @return True if the given information is stored in the merkle tree, false otherwise
     function isAccountBalanceCorrect(
-        uint256 merkleRoot,
-        uint24  accountID,
-        uint16  tokenID,
-        uint256 pubKeyX,
-        uint256 pubKeyY,
-        uint32  nonce,
-        uint96  balance,
-        uint256 tradeHistoryRoot,
-        uint256[30] calldata accountMerkleProof,
-        uint256[12] calldata balanceMerkleProof
+        uint     merkleRoot,
+        uint24   accountID,
+        uint16   tokenID,
+        uint     pubKeyX,
+        uint     pubKeyY,
+        uint32   nonce,
+        uint96   balance,
+        uint     tradeHistoryRoot,
+        uint[30] calldata accountMerkleProof,
+        uint[12] calldata balanceMerkleProof
         )
         external
         pure
@@ -526,8 +526,8 @@ contract IExchange
     /// @param blockIndices The 0-based index of the blocks to be verified with the given proofs
     /// @param proofs The ZK proof for all blockIndices (proofs.length % 8 == 0).
     function verifyBlocks(
-        uint[]    calldata blockIndices,
-        uint256[] calldata proofs
+        uint[] calldata blockIndices,
+        uint[] calldata proofs
         )
         external;
 
@@ -580,7 +580,7 @@ contract IExchange
         view
         returns (
           bytes32 accumulatedHash,
-          uint256 accumulatedFee,
+          uint    accumulatedFee,
           uint32  timestamp
         );
 
@@ -699,7 +699,7 @@ contract IExchange
         view
         returns (
             bytes32 accumulatedHash,
-            uint256 accumulatedFee,
+            uint    accumulatedFee,
             uint32  timestamp
         );
 
@@ -768,14 +768,14 @@ contract IExchange
     ///                      token for the account. The deepest hash in the tree is the
     ///                      1st element of the array.
     function withdrawFromMerkleTree(
-        address token,
-        uint    pubKeyX,
-        uint    pubKeyY,
-        uint32  nonce,
-        uint96  balance,
-        uint256 tradeHistoryRoot,
-        uint256[30] calldata accountMerkleProof,
-        uint256[12] calldata balanceMerkleProof
+        address  token,
+        uint     pubKeyX,
+        uint     pubKeyY,
+        uint32   nonce,
+        uint96   balance,
+        uint     tradeHistoryRoot,
+        uint[30] calldata accountMerkleProof,
+        uint[12] calldata balanceMerkleProof
         )
         external;
 
@@ -802,15 +802,15 @@ contract IExchange
     ///                      token for the account. The deepest hash in the tree is the
     ///                      1st element of the array.
     function withdrawFromMerkleTreeFor(
-        address owner,
-        address token,
-        uint    pubKeyX,
-        uint    pubKeyY,
-        uint32  nonce,
-        uint96  balance,
-        uint256 tradeHistoryRoot,
-        uint256[30] calldata accountMerkleProof,
-        uint256[12] calldata balanceMerkleProof
+        address  owner,
+        address  token,
+        uint     pubKeyX,
+        uint     pubKeyY,
+        uint32   nonce,
+        uint96   balance,
+        uint     tradeHistoryRoot,
+        uint[30] calldata accountMerkleProof,
+        uint[12] calldata balanceMerkleProof
         )
         external;
 
@@ -868,7 +868,7 @@ contract IExchange
     /// @param  feeRecipient The address that receives the block fee
     /// @return feeAmount The amount of ETH earned in the block and sent to the operator
     function withdrawBlockFee(
-        uint blockIdx,
+        uint    blockIdx,
         address payable feeRecipient
         )
         external
